+++ cmd/state-tray/main_linux.go
@@ -1,5 +1,5 @@
 package main
 
 import (
-	_ "github.com/ActiveState/cli/cmd/state-tray/internal/libc-compatwrap"
+	_ "github.com/ActiveState/cli/internal/libc-compatwrap"
 )