+++ internal/runners/update/update.go
@@ -53,6 +53,9 @@ func (u *Update) Run(params *Params) error {
 	}
 	up, err := m.InitiateDeferredUpdate(channel, "")
 	if err != nil {
+		if channel == constants.BetaBranch || channel == constants.ReleaseBranch {
+			return locale.NewInputError("err_unsupported_update", "The current version of the State Tool cannot update to the target channel {{.V0}}.  You can still run the installation one-liners to update the State Tool. See {{.V1}} for details.", channel, "https://www.activestate.com/products/platform/state-tool/")
+		}
 		return locale.WrapError(err, "err_update_initiate", "Failed to initiate update.")
 	}
 	if up.Channel == "" && up.Version == "" {