+++ pkg/platform/model/svc.go
@@ -22,7 +22,11 @@ func NewSvcModel(ctx context.Context, cfg *config.Instance) (*SvcModel, error) {
 	if err != nil {
 		return nil, errs.Wrap(err, "Could not initialize svc client")
 	}
-	return &SvcModel{ctx, client}, nil
+	return NewSvcModelWithClient(ctx, client), nil
+}
+
+func NewSvcModelWithClient(ctx context.Context, client *gqlclient.Client) *SvcModel {
+	return &SvcModel{ctx, client}
 }
 
 func (m *SvcModel) StateVersion() (*graph.Version, error) {
@@ -50,4 +54,13 @@ func (m *SvcModel) InitiateDeferredUpdate(channel, version string) (*graph.Defer
 		return nil, locale.WrapError(err, "err_svc_updaterequest", "Error updating to version {{.V0}} at channel {{.V1}}: {{.V2}}", version, channel, errs.Join(err, ": ").Error())
 	}
 	return &u.DeferredUpdate, nil
+}
+
+func (m *SvcModel) CheckUpdate() (*graph.AvailableUpdate, error) {
+	r := request.NewAvailableUpdate()
+	u := graph.AvailableUpdateResponse{}
+	if err := m.client.Run(r, &u); err != nil {
+		return nil, errs.Wrap(err, "Error checking if update is available.")
+	}
+	return &u.AvailableUpdate, nil
 }